% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_curves.R
\docType{data}
\name{example_settings}
\alias{example_settings}
\title{get_derivatives function}
\format{An object of class \code{list} of length 48.}
\usage{
example_settings
}
\arguments{
\item{data}{- вектор со значениями флуорисценции}
}
\description{
Функция считает первую и вторую производную от кривой флуорисценции
}
\examples{
get_derivatives(curve)
}
\keyword{derivative}
\keyword{estimation}

