% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_curve.R
\name{get_curve}
\alias{get_curve}
\title{get_curve function}
\usage{
get_curve(data, channel, curve)
}
\arguments{
\item{data}{- вектор со значениями флуорисценции}
}
\description{
Функция считает первую и вторую производную от кривой флуорисценции
}
\examples{
get_derivatives(curve)
}
\keyword{derivative}
\keyword{estimation}

