% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_curves.R
\name{get_curves}
\alias{get_curves}
\title{get_derivatives function}
\usage{
get_curves(data, settings = NULL)
}
\arguments{
\item{data}{- вектор со значениями флуорисценции}
}
\description{
Функция считает первую и вторую производную от кривой флуорисценции
}
\examples{
get_derivatives(curve)
}
\keyword{derivative}
\keyword{estimation}

