% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_derivatives.R
\name{get_derivatives}
\alias{get_derivatives}
\title{get_derivatives function}
\usage{
get_derivatives(data)
}
\arguments{
\item{data}{- вектор со значениями флуорисценции}
}
\description{
Функция считает первую и вторую производную от кривой флуорисценции
}
\examples{
get_derivatives(curve)
}
\keyword{derivative}
\keyword{estimation}

