% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dt.R
\name{read_dt}
\alias{read_dt}
\title{read_dt function}
\usage{
read_dt(file_path)
}
\arguments{
\item{file_path}{- расположение файла}
}
\description{
Функция парсит файлы приборов ДТ форматов .r48, .r96
}
\examples{
read_dt("../../pcr_read.r48")
}
\keyword{DT}
\keyword{parse,}

