% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dt.R
\name{read_dt}
\alias{read_dt}
\title{A Cat Function}
\usage{
read_dt(file_path)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

